simple http interface
This adds a small HTTP interface to each node, allowing for retrieval
and placement of keys. The interface maintains parity with the internal
node interface and as such normal queries may be performed. For example,
a node running on localhost:3000 might be queried over HTTP with curl
like so:

    $ curl -XPUT localhost:3000 -H 'content-type: application/json' \
           -d '{"k": "foo", "v": {}}'

which retrieves the full value for the key "foo". Note that search and
aggregate queries also work here. For example:

    $ curl -XPUT localhost:3000 -H 'content-type: application/json' \
           -d '{"k": "foo%", "v": {}}'

which performs a search query that is equivalent to the regex "foo.*".

Results are returned in an array. This facilitates search queries which
may return multiple responses.