add merkle tree implementation
This includes an implementation of a Merkle tree optimized for handling
sorted collections of key-values that may be larger than RAM. The
particulars of this implementation and a description of its mechanics are
as follows.

    Example input:

    [[:a 1] [:b 2] [:c 3]]

    Depiction of example output tree:

               o :a :c
              / \
       :a :b o   o :c :c
            / \
     :a :a o   o :b :b

Given a sorted sequence of "tuple" sequences which contain first the key
and then the value, a tree is constructed from the bottom up. The head
of this sequence becomes the left-most node in the tree. First this node
is passed to a recursive procedure which uses a vector to store
intermediary subtrees. The index that a subtree occupies is relative to
its height. Nodes are generated from the sorted sequence and passed in
to this procedure, always beginning with the 0th index. When an index is
already occupied, we join the passed in node with the node occupying
that index and then recursively call the procedure, incrementing the
index. When we have processed the entire sequence we are left with a
sequence of subtrees.

With the result of the above procedure we can take this sequence and
reduce it with a node-joining function to produce a complete tree. Care
must be taken to reverse this sequence before processing its reduction.
This is because the left-most subtrees are the deepest subtrees and
therefore should be merged with smaller subtrees on their right side.
Intuitively we can think of this as joining the shallowest subtrees (the
right-most indices) on the right side of the deepest trees. The result
is a tree which is preferenced towards its left side.

This implementation is intended to be used in our active anti-entropy
protocol. Quanta nodes will periodically exchange Merkle tree root
hashes. When hashes mismatch, we an exchange will take place between the
disagreeing nodes that which will result in the missing key ranges. With
this information a node can request these values from the cluster and
thereby heal its local state.