fix singleton node http updates
This fixes a bug where a singleton node, i.e. a node with an empty peer
list, would fail to update a vector via the HTTP API. Previously we
produced a lazy sequence which when realized mapped the
`handler/forward!` function over an incoming message. However this
sequence was nested within the scope of a `when-let` form predicated on a
non-empty peer list. Put another way: when a node didn't have a peer
list, it never realized the forwards for a message resulting lost
updates. Here we fix this by realizing the sequence before checking for
a peer list.